refactor: customEventManager util로 분리
