+++ apps/enfpy-native/src/components/web-view/CommonWebView.tsx
@@ -19,6 +19,7 @@ import {
 import {createRequestMessageHandler} from '../../contexts/web-bridge/handlers';
 import useConfig from '../../hooks/useConfig';
 import {Fetcher, buildFetcher} from '@vrew/modules/app-bridge/buildFetcher';
+import {buildCustomEventManager} from '../../utils/customEvent';
 
 export interface CommonWebViewProps extends WebViewProps {
   source: WebViewSourceUri;
@@ -33,34 +34,6 @@ const isValidUrl = (url: string) => {
   // return regExp.test(url);
 
   return url.startsWith('http');
-};
-
-const buildCustomEventManager = () => {
-  let eventMap: Record<string, ((event: any) => void)[]> = {};
-
-  const addEventListener = (eventName: string, callbackFn: any) => {
-    const handlers = eventMap[eventName] || [];
-    eventMap[eventName] = handlers.concat(callbackFn);
-  };
-
-  const emit = (eventName: string, event: any) => {
-    const handlers = eventMap[eventName] || [];
-
-    handlers.forEach(handler => {
-      handler(event);
-    });
-  };
-
-  const removeEventListener = (eventName: string, eventListener: any) => {
-    const handlers = eventMap[eventName] || [];
-    eventMap[eventName] = handlers.filter(handler => handler !== eventListener);
-  };
-
-  return {
-    removeEventListener,
-    addEventListener,
-    emit,
-  };
 };
 
 export interface CommonWebViewRef {
